Use guint for array size everywhere
Using a GPtrArray for underlying storage means that our length is
restricted to what a guint can represent, so there's no point in
pretending we can do better.

Although it would kind of be nice to use a long and have -1 as an
error value. Nobody uses 32-bit anymore, right?

+++ gcollection/gc-array-list.h
@@ -40,9 +40,9 @@ GcArrayList     *gc_array_list_new_full               (GType          element_ty
 GPtrArray       *gc_array_list_get_ptr_array          (GcArrayList *array);
 
 gpointer         gc_array_list_get                    (GcArrayList *array,
-                                                       gsize        index);
+                                                       guint        index);
 
-gsize            gc_array_list_get_size               (GcArrayList *array);
+guint            gc_array_list_get_size               (GcArrayList *array);
 
 gboolean         gc_array_list_empty                  (GcArrayList *array);
